fix: pytest error -> warning
